Add SingleRegion option for testing stackset rollouts
Stacksets that roll out to many regions take a long time to complete. AWS recommends
rolling out to a single region to test a change to a stackset before rolling out to all target regions.

This is now supported with a SingleRegion argument. To test an inital rollout:

  make main-rollout SingleRegion=1

To test a stackset change:

  make stacksets-update SingleRegion=1

Note that this does not affect the Validate tags so if you don't follow up with a full rollout
the Config Rule that validates stackset deployments will flag the stackset as non-compliant.